commit -a

+++ moduler.py
@@ -22,8 +22,8 @@ def load_etf_data(type, code) :
     tmp = tmp.loc[tmp['etf_code'] == code, :]
     tmp = tmp.drop('etf_code', axis = 1)
     tmp.columns = ['종목코드', '종목명', '보유량', '평가금액', '비중']
-    tmp['보유량'] = tmp['보유량'].astype(int)
-    tmp['평가금액'] = tmp['평가금액'].astype(int)
+    tmp['보유량'] = tmp['보유량'].astype(float)
+    tmp['평가금액'] = tmp['평가금액'].astype(float)
     tmp['비중'] = tmp['비중'].astype(float)
 
     return tmp
